==> preparations/03_calculated_aspects_ratios.py <==

p_aspects={
'sun':{
'moon':[0.0428,0.0328,0.0905,0.0897,0.0428,0.2985],
'mercury':[0.2298,0.0181,0,0,0,0.2479],
'venus':[0.1205,0.1108,0,0,0,0.2313],
'mars':[0.0724,0.0533,0.0886,0.0532,0.0148,0.2823],
'jupiter':[0.0518,0.0387,0.0888,0.0803,0.0350,0.2946],
'saturn':[0.0479,0.0363,0.0889,0.0843,0.0391,0.2965],
'uranus':[0.0456,0.0347,0.0888,0.0865,0.0414,0.2970],
'neptune':[0.0448,0.0342,0.0888,0.0873,0.0420,0.2972],
'pluto':[0.0447,0.0341,0.0889,0.0874,0.0424,0.2975],
'tnode':[0.0458,0.0341,0.0843,0.0865,0.0458,0.2965],
'mnode':[0.0434,0.0333,0.0889,0.0888,0.0435,0.2978],
'lilit':[0.0434,0.0332,0.0889,0.0888,0.0434,0.2978],
'chiron':[0.0458,0.0354,0.0888,0.0856,0.0408,0.2964]
},
'moon':{
'mercury':[0.0428,0.0332,0.0902,0.0895,0.0428,0.2985],
'venus':[0.0431,0.0331,0.0897,0.0893,0.0430,0.2982],
'mars':[0.0437,0.0332,0.0889,0.0888,0.0431,0.2976],
'jupiter':[0.0435,0.0332,0.0889,0.0885,0.0433,0.2974],
'saturn':[0.0435,0.0331,0.0886,0.0888,0.0434,0.2974],
'uranus':[0.0434,0.0332,0.0889,0.0889,0.0438,0.2981],
'neptune':[0.0435,0.0332,0.0888,0.0887,0.0435,0.2977],
'pluto':[0.0433,0.0333,0.0889,0.0886,0.0436,0.2977],
'tnode':[0.0438,0.0334,0.0881,0.0884,0.0437,0.2974],
'mnode':[0.0436,0.0333,0.0884,0.0887,0.0436,0.2977],
'lilit':[0.0484,0.0363,0.0884,0.0837,0.0389,0.2957],
'chiron':[0.0426,0.0332,0.0887,0.0887,0.0440,0.2972]
},
'mercury':{
'venus':[0.1262,0.1040,0,0,0,0.2302],
'mars':[0.0731,0.0508,0.0888,0.0570,0.0144,0.2841],
'jupiter':[0.0519,0.0386,0.0886,0.0808,0.0354,0.2953],
'saturn':[0.0479,0.0367,0.0891,0.0842,0.0389,0.2968],
'uranus':[0.0453,0.0346,0.0890,0.0870,0.0417,0.2976],
'neptune':[0.0450,0.0342,0.0889,0.0874,0.0419,0.2974],
'pluto':[0.0438,0.0335,0.0888,0.0884,0.0430,0.2975],
'tnode':[0.0456,0.0336,0.0850,0.0868,0.0452,0.2963],
'mnode':[0.0437,0.0331,0.0888,0.0888,0.0433,0.2976],
'lilit':[0.0428,0.0338,0.0895,0.0885,0.0433,0.2979],
'chiron':[0.0450,0.0353,0.0892,0.0865,0.0412,0.2973]
},
'venus':{
'mars':[0.0697,0.0502,0.0889,0.0635,0.0168,0.2891],
'jupiter':[0.0506,0.0383,0.0890,0.0816,0.0365,0.2960],
'saturn':[0.0466,0.0358,0.0889,0.0851,0.0400,0.2963],
'uranus':[0.0449,0.0349,0.0881,0.0870,0.0419,0.2968],
'neptune':[0.0443,0.0343,0.0891,0.0875,0.0426,0.2979],
'pluto':[0.0443,0.0342,0.0888,0.0875,0.0427,0.2975],
'tnode':[0.0443,0.0338,0.0867,0.0877,0.0446,0.2972],
'mnode':[0.0431,0.0332,0.0889,0.0888,0.0436,0.2977],
'lilit':[0.0443,0.0337,0.0890,0.0885,0.0427,0.2981],
'chiron':[0.0467,0.0345,0.0885,0.0835,0.0414,0.2945]
},
'mars':{
'jupiter':[0.0444,0.0337,0.0944,0.0739,0.0466,0.2930],
'saturn':[0.0450,0.0338,0.0885,0.0871,0.0430,0.2975],
'uranus':[0.0439,0.0331,0.0886,0.0874,0.0437,0.2967],
'neptune':[0.0438,0.0337,0.0883,0.0888,0.0439,0.2985],
'pluto':[0.0433,0.0332,0.0884,0.0881,0.0439,0.2970],
'tnode':[0.0453,0.0327,0.0913,0.0874,0.0456,0.3024],
'mnode':[0.0447,0.0329,0.0915,0.0874,0.0456,0.3021],
'lilit':[0.0433,0.0331,0.0887,0.0874,0.0443,0.2969],
'chiron':[0.0400,0.0349,0.0853,0.0876,0.0466,0.2984]
},
'jupiter':{
'saturn':[0.0446,0.0350,0.0909,0.0872,0.0419,0.2996],
'uranus':[0.0400,0.0341,0.0878,0.0886,0.0443,0.2987],
'neptune':[0.0440,0.0332,0.0886,0.0887,0.0441,0.2985],
'pluto':[0.0406,0.0314,0.0890,0.0913,0.0441,0.2964],
'tnode':[0.0433,0.0338,0.0909,0.0875,0.0441,0.2995],
'mnode':[0.0431,0.0335,0.0895,0.0882,0.0429,0.2972],
'lilit':[0.0462,0.0365,0.0884,0.0891,0.0438,0.3039],
'chiron':[0.0436,0.0346,0.0892,0.0905,0.0474,0.3052]
},
'saturn':{
'uranus':[0.0438,0.0343,0.0905,0.0910,0.0414,0.3010],
'neptune':[0.0464,0.0399,0.0887,0.0899,0.0436,0.3084],
'pluto':[0.0412,0.0310,0.0894,0.0916,0.0434,0.2967],
'tnode':[0.0442,0.0344,0.0885,0.0839,0.0422,0.2932],
'mnode':[0.0444,0.0360,0.0885,0.0828,0.0426,0.2944],
'lilit':[0.0421,0.0299,0.0884,0.0888,0.0447,0.2939],
'chiron':[0.0223,0.0368,0.1032,0.0732,0.0490,0.2845]
},
'uranus':{
'neptune':[0.0471,0.0360,0.0879,0.0845,0.0366,0.2921],
'pluto':[0.0428,0.0335,0.1000,0.0863,0.0367,0.2993],
'tnode':[0.0438,0.0331,0.0895,0.0908,0.0447,0.3020],
'mnode':[0.0439,0.0328,0.0893,0.0901,0.0445,0.3006],
'lilit':[0.0429,0.0329,0.0885,0.0889,0.0436,0.2968],
'chiron':[0.0543,0.0558,0.1147,0.0447,0.0506,0.3002]
},
'neptune':{
'pluto':[0.0275,0.0241,0.0467,0.1593,0.0206,0.2781],
'tnode':[0.0434,0.0333,0.0892,0.0892,0.0427,0.2977],
'mnode':[0.0434,0.0333,0.0891,0.0891,0.0430,0.2979],
'lilit':[0.0438,0.0335,0.0888,0.0881,0.0426,0.2967],
'chiron':[0.0505,0.0283,0.0832,0.1006,0.0399,0.3025]
},
'pluto':{
'tnode':[0.0433,0.0336,0.0895,0.0872,0.0430,0.2967],
'mnode':[0.0434,0.0333,0.0895,0.0878,0.0430,0.2970],
'lilit':[0.0432,0.0334,0.0890,0.0878,0.0434,0.2967],
'chiron':[0.0418,0.0335,0.0906,0.0710,0.0575,0.2944]
},
'tnode':{
'mnode':[0.9776,0.0001,0,0,0,0.9777],
'lilit':[0.0435,0.0334,0.0885,0.0882,0.0436,0.2972],
'chiron':[0.0422,0.0327,0.0875,0.0906,0.0433,0.2964]
},
'mnode':{
'lilit':[0.0436,0.0332,0.0881,0.0890,0.0437,0.2976],
'chiron':[0.0422,0.0337,0.0883,0.0898,0.0435,0.2974]
},
'lilit':{
'chiron':[0.0433,0.0318,0.0891,0.0890,0.0435,0.2966]
}
}
